AB#12505-Modified search component as parent for advanced search and simplified search and done the necessary changes
