+++ src/app/features/fss-search/fss-popular-search-batches/fss-popular-search-batches.component.html
@@ -1,5 +1,5 @@
 <h3>Popular searches</h3>
-<div *ngFor="let popularSearch of popularSearches" (click)="getPopularSearch(popularSearch)" class="SearchBatch">
+<div *ngFor="let popularSearch of popularSearches" (click)="onPopularSearchClick(popularSearch)" class="SearchBatch">
     <p *ngIf="popularSearch.isActive === true" class='popular-search-para'>
         {{popularSearch.displayName}} 
         <button class="popular-search fa fa-chevron-right" aria-label="Popular Search Row"></button>
